+++ builder/webpack.config.js
@@ -43,7 +43,13 @@ module.exports = {
   },
 
   postcss: () => {
-    return [autoprefixer('last 2 version')];
+    return [autoprefixer({
+      browsers: [
+        'last 2 versions',
+        'iOS >= 8',
+        'Safari >= 8',
+      ],
+    })];
   },
 
   plugins: [