+++ components/index.js
@@ -1,7 +1,9 @@
 import Button from './button/button';
 import Calendar from './calendar/calendar';
+import Checkbox from './checkbox/checkbox';
 import Collapse from './collapse/collapse';
 import CollapseItem from './collapse/collapseItem';
+import DropDown from './dropDown/dropDown';
 import List from './list/list';
 import Modal from './modal/modal';
 import Price from './price/price';
@@ -11,8 +13,10 @@ import Spinner from './spinner/spinner';
 export default {
   Button,
   Calendar,
+  Checkbox,
   Collapse,
   CollapseItem,
+  DropDown,
   List,
   Modal,
   Price,